+++ cmd/render.go
@@ -2,17 +2,20 @@ package cmd
 
 import (
 	"fmt"
+	"sync"
 
 	"github.com/dhth/ecsv/internal/aws"
 	"github.com/dhth/ecsv/internal/types"
 	"github.com/dhth/ecsv/internal/ui"
 )
 
-func render(systems []types.System, config ui.Config, awsConfigs map[string]aws.Config) error {
+func render(systems []types.System, config ui.Config, awsConfigs map[string]aws.Config, maxConcFetches int) error {
 	results := make(map[string]map[string]types.SystemResult)
 	resultChannel := make(chan types.SystemResult)
 
-	counter := 0
+	semaphore := make(chan struct{}, maxConcFetches)
+	var wg sync.WaitGroup
+
 	for _, s := range systems {
 		awsConfig := awsConfigs[s.AWSConfigKey()]
 		if results[s.Key] == nil {
@@ -28,14 +31,25 @@ func render(systems []types.System, config ui.Config, awsConfigs map[string]aws.
 			}
 			continue
 		}
+
+		wg.Add(1)
+
 		go func(system types.System) {
+			defer wg.Done()
+			semaphore <- struct{}{}
+			defer func() {
+				<-semaphore
+			}()
 			resultChannel <- aws.FetchSystemVersion(system, awsConfig)
 		}(s)
-		counter++
 	}
 
-	for range counter {
-		r := <-resultChannel
+	go func() {
+		wg.Wait()
+		close(resultChannel)
+	}()
+
+	for r := range resultChannel {
 		results[r.SystemKey][r.Env] = r
 	}